feat: 🎸 Move createVenue into settlements namespace
sdk.currentIdentity.createVenue -> sdk.settlements.createVenue

BREAKING CHANGE: 🧨 createVenue moved from currentIdentity to settlements
