+++ src/CurrentIdentity.ts
@@ -2,8 +2,6 @@ import {
   Context,
   createSecurityToken,
   CreateSecurityTokenWithTickerParams,
-  createVenue,
-  CreateVenueParams,
   inviteAccount,
   InviteAccountParams,
   modifySignerPermissions,
@@ -15,7 +13,6 @@ import {
   SecurityToken,
   TickerReservation,
   toggleFreezeSecondaryKeys,
-  Venue,
 } from '~/internal';
 import { NoArgsProcedureMethod, PermissionType, ProcedureMethod, Signer } from '~/types';
 import { createProcedureMethod } from '~/utils/internal';
@@ -70,10 +67,6 @@ export class CurrentIdentity {
       { getProcedureAndArgs: args => [inviteAccount, { ...args }] },
       context
     );
-    this.createVenue = createProcedureMethod(
-      { getProcedureAndArgs: args => [createVenue, args] },
-      context
-    );
     this.freezeSecondaryKeys = createProcedureMethod(
       { getProcedureAndArgs: () => [toggleFreezeSecondaryKeys, { freeze: true }], voidArgs: true },
       context
@@ -119,11 +112,6 @@ export class CurrentIdentity {
    *   fetch its pending Authorization Requests by calling `authorizations.getReceived`
    */
   public inviteAccount: ProcedureMethod<InviteAccountParams, void>;
-
-  /**
-   * Create a Venue under the ownership of the Current Identity
-   */
-  public createVenue: ProcedureMethod<CreateVenueParams, Venue>;
 
   /**
    * Freeze all the secondary keys in the Current Identity. This means revoking their permission to perform any operation on the blockchain and freezing their funds until the keys are unfrozen via [[unfreezeSecondaryKeys]]
